+++ src/qcommon/cm_load.c
@@ -732,7 +732,7 @@ void CM_LoadMap( const qchar *name, qbool clientload, qint *checksum ) {
 #if !defined(BSPC)
 	cm_noAreas = Cvar_GetAndDescribe("cm_noAreas", "0", CVAR_CHEAT, "Do not use areaportals, all areas are connected.");
 	cm_noCurves = Cvar_GetAndDescribe("cm_noCurves", "0", CVAR_CHEAT, "Do not collide against curves.");
-	cm_playerCurveClip = Cvar_GetAndDescribe("cm_playerCurveClip", "1", CVAR_ARCHIVE | CVAR_CHEAT, "Collide against player curves.");
+	cm_playerCurveClip = Cvar_GetAndDescribe("cm_playerCurveClip", "1", CVAR_ARCHIVE_ND | CVAR_CHEAT, "Collide against player curves.");
 #endif
 	Com_DPrintf( "CM_LoadMap( %s, %i )\n", name, clientload );